test: consolidate Z2M E2E tests into test_zigbee2mqtt.py, remove superseded unit tests
Entire-Checkpoint: 37a1ad0f632d
