--- tests/providers/zigbee2mqtt/__init__.py
@@ -1,1 +1,0 @@
-"""Zigbee2MQTT provider tests."""
